Step5 - 자동차 경주(리팩토링) (#4471)
* refactor: 내부 인스턴스 변수들을 일급 컬렉션으로 분리

* refactor: 패키지 구조 변경

* refactor: 원시값 포장 객체를 비교 가능하게 변경

* test: 빠진 도메인 테스트 추가

+++ src/main/java/racingcar/model/dto/CarDto.java
@@ -1,6 +1,6 @@
 package racingcar.model.dto;
 
-import racingcar.model.Car;
+import racingcar.model.car.Car;
 
 public class CarDto {
     private final int distance;
@@ -12,7 +12,7 @@ public class CarDto {
     }
 
     public static CarDto from(Car car) {
-        return new CarDto(car.name(), car.distance());
+        return new CarDto(car.name(), car.distance().intValue());
     }
 
     public String name() {
